# HDDS-12345. Share cluster in filesystem tests