## What changes were proposed in this pull request?

Speed up the following filesystem tests by adding to `NonHATests`:

- `TestOzoneFSBucketLayout`
- `TestOzoneFSInputStream`
- `TestOzoneFSWithObjectStoreCreate`
- `TestOzoneFileSystemMetrics`
- `TestOzoneFileSystemMissingParent`

Fix `TestContainerOperations` to leave datanode in its original state, not in `DECOMMISSIONING`.

https://issues.apache.org/jira/browse/HDDS-12345

## How was this patch tested?

Before:

```
Tests run: 5, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 35.94 s -- in org.apache.hadoop.fs.ozone.TestOzoneFSBucketLayout
Tests run: 13, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 51.02 s -- in org.apache.hadoop.fs.ozone.TestOzoneFSInputStream
Tests run: 6, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 37.97 s -- in org.apache.hadoop.fs.ozone.TestOzoneFSWithObjectStoreCreate
Tests run: 3, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 40.05 s -- in org.apache.hadoop.fs.ozone.TestOzoneFileSystemMetrics
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 36.49 s -- in org.apache.hadoop.fs.ozone.TestOzoneFileSystemMissingParent
```

After:

```
Tests run: 5, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.219 s -- in org.apache.ozone.test.NonHATests$OzoneFSBucketLayout
Tests run: 13, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 13.02 s -- in org.apache.ozone.test.NonHATests$OzoneFSInputStream
Tests run: 6, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.860 s -- in org.apache.ozone.test.NonHATests$OzoneFSWithObjectStoreCreate
Tests run: 3, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 4.798 s -- in org.apache.ozone.test.NonHATests$OzoneFileSystemMetrics
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.164 s -- in org.apache.ozone.test.NonHATests$OzoneFileSystemMissingParent
```

https://github.com/adoroszlai/ozone/actions/runs/13496989456